% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGeneSelection.R
\name{runGeneSelection}
\alias{runGeneSelection}
\title{RNA-seq Differential Expression Analysis Shiny App}
\value{
User Interface for RNA-seq Differential Expression Analysis from scratch
}
\description{
The Shiny application provides a user-friendly platform for performing detailed differential gene expression analyses for RNA-seq Data.
It provides a comprehensive set of tools for processing raw FASTQ files,
quality assessment, filtering, trimming, alignment to reference genomes,
read counting, and identify the differentiated expressed genes.
}
\examples{
## You can simply be run using and input files into the shiny app
runGeneSelection()
}
